+++ packages/docusaurus-theme-classic/src/theme/DocSidebarItem/Html.module.css
@@ -11,7 +11,3 @@
       var(--ifm-menu-link-padding-horizontal);
   }
 }
-
-.menuExternalLink {
-  align-items: center;
-}
